backend/requirements.txt
fastapi
uvicorn
pdfplumber
pandas
sentence-transformers
faiss-cpu
numpy
openpyxl
python-multipart
